+++ src/new/brute-force/main-bf_par_articulation.cpp
@@ -56,6 +56,7 @@ extern int num_threads;
 int output = 0;
 bool best = false;
 int matrix_t=0;
+bool nolb = false;
 
 /// @brief  The main method
 int main(int argc, char** argv){