+++ src/new/brute-force/main-bf_par_cycles_m1.cpp
@@ -56,8 +56,8 @@ int max_induced_cycles;
 
 int output = 0;
 bool best = false;
-
 int matrix_t=0;
+bool nolb = false;
 
 /// @brief  The main method
 int main(int argc, char** argv){