+++ src/new/brute-force/main-bf_par_threads.cpp
@@ -55,8 +55,8 @@ extern int num_threads;
 
 int output = 0;
 bool best = false;
-
 int matrix_t=0;
+bool nolb = false;
 
 /// @brief  The main method
 int main(int argc, char** argv){